refactor(validate): consolidate body-flag helpers into internal/validate
internal/validate/ already exists and its doc.go explicitly
declares itself the anchor for "future non-path validators".
A separate internal/cli/add/core/validate/ duplicated that
philosophy with a tiny pure-function helper. Fold the helper
into internal/validate/ and delete the duplicate.

Changes:
- internal/validate/bodyflags.go — new file with BodyFlags
  and RejectPlaceholder (identical to the deleted helpers).
- internal/validate/bodyflags_test.go — tests moved over.
- internal/validate/doc.go — adds a "CLI Body-Flag Validation"
  section alongside the existing path validators.
- internal/cli/decision/cmd/add/cmd.go and
  internal/cli/learning/cmd/add/cmd.go — import
  internal/validate instead of the old path.
- internal/cli/add/core/validate/ — deleted.

Spec: specs/skill-surface-polish.md
