+++ internal/cli/learning/cmd/add/cmd.go
@@ -10,10 +10,10 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/ActiveMemory/ctx/internal/cli/add/core/build"
-	"github.com/ActiveMemory/ctx/internal/cli/add/core/validate"
 	"github.com/ActiveMemory/ctx/internal/config/embed/cmd"
 	"github.com/ActiveMemory/ctx/internal/config/entry"
 	cFlag "github.com/ActiveMemory/ctx/internal/config/flag"
+	"github.com/ActiveMemory/ctx/internal/validate"
 )
 
 // Cmd returns the "ctx learning add" subcommand.
